attempt to reduce amount of generated testCases per testSuite, by seperating some code in helper functions and creating four testSuite files out of one (reduce results by internal used generator). This is required to not reach CIs timeout limits. This commit might get reverted in case it does not solve our problem. In case it doesn't solve our problem, we need to extend the functionality of testing.js and test-utils.js to be able to not only break down testSuites into seperate Buckets, but add Bucket support for testSuites itself (if required). Currently the first and easier approach is taken, as it seems that a lot of adjustments throughout the code-base is required to implement bucket testSuite support - as currently the overall mechanism is based on fileNames/testSuites itself and not list of testCases.
